Make TinyBasic compile and work
Note that you still need to split the file up to get it compile with zasm.
Or you can get the pre-split version (with some more features, like lowercase variables/keywords and IN/OUT/PEEK/POKE/SYS) at 
https://github.com/schierlm/CollapseOS-Web-Emulator/tree/master/basic

Or watch it live on https://schierlm.github.io/CollapseOS-Web-Emulator/emul1.html (just run `basic`).

Fixes: 
- Remove leading zeroes for 16-bit constants (0x0FFFF -> 0xFFFF)
- Add hardware specific part (include "user.h")
- Enable output switch OCSW by default
- do not block when checking for input (to break loops)
- Eliminate forward references in `.db` directive
- Fix stack imbalance in `TX_RDY`
- Remove spaces around arithmetic + operator
- Eliminate multiple `.org` statements


And last but not least, remove a logic error in the ZASM workaround for the jump address in the command tables.

As the 0x80 marker is on the second byte of the jump address, it can happen that the first byte also is larger than 0x80,
which would make the marker detected one byte too early.

I. e. "LIST" 0xCC 0xDD would be detected as "LIS" "T" 0xCC and jump to some random place.

Also, due to empty list entries, there may be up to 2 consecutive jumps, of which both may
have two bytes larger than 0x80.

Fix this by introducing new `CKJ` function that checks up to 5 bytes and determines from the values where the real break is.
Utilize the fact that the byte *after* the jump will be less than 0x80, and the second
byte of the jump will always be greater.

(We also need to add an extra 0x00 at the very end of all the tables to make this work, 
so that we can be sure that the byte after the last jump will also be less than 0x80).